Move WorkflowMixin to workflows submodule
